Support for linux bootstrapping.

+++ dewey/bin/bootstrap_dewey.sh
@@ -1,11 +1,20 @@
 cd ~
 docopt-completion hey_dewey --manual-bash > /dev/null
-mv hey_dewey.sh .dewey_autocomplete.sh
-
-echo "" >> .dewey_autocomplete.sh
-echo "complete -F _hey_dewey hey_dewey" >> .dewey_autocomplete.sh
-echo "complete -F _hey_dewey dewey" >> .dewey_autocomplete.sh
-echo "complete -F _hey_dewey d" >> .dewey_autocomplete.sh
+if [ ! -f ./hey_dewey.sh ]; then
+    sudo docopt-completion hey_dewey --manual-bash > /dev/null
+    cd /etc/bash_completion.d/
+    sudo mv hey_dewey.sh .dewey_autocomplete.sh
+    echo "" | sudo tee .dewey_autocomplete.sh
+    echo "complete -F _hey_dewey hey_dewey" | sudo tee .dewey_autocomplete.sh
+    echo "complete -F _hey_dewey dewey" | sudo tee .dewey_autocomplete.sh
+    echo "complete -F _hey_dewey d" | sudo tee .dewey_autocomplete.sh
+else
+    mv hey_dewey.sh .dewey_autocomplete.sh
+    echo "" >> .dewey_autocomplete.sh
+    echo "complete -F _hey_dewey hey_dewey" >> .dewey_autocomplete.sh
+    echo "complete -F _hey_dewey dewey" >> .dewey_autocomplete.sh
+    echo "complete -F _hey_dewey d" >> .dewey_autocomplete.sh
+fi;
 
 source .dewey_autocomplete.sh
